clk: vt8500: Remove duplicate divisor adjustment in vt8500_dclk_set_rate()
The calculated divisor is already adjusted in vt8500_dclk_round_rate(), and
may be re-adjusted in set_rate due to round errors. clk_set_rate() calls
vt8500_dclk_round_rate() and adjusts the requested rate to a rounded value
that will work. We don't need to test whether the divisor is too low because
the adjustment is already made.

Signed-off-by: Tony Prisk <linux@prisktech.co.nz>