dts: vt8500: Add framebuffer support for APC8750
This patch enables framebuffer support for the APC8750 via the VGA output.
A default display config of 1024x768x60 is provided, which is the default
uboot resolution.

A change to the binding is required to allow a clock to be passed to config
the pixel clock. This binding needs to be changed properly at a later time.

Signed-off-by: Tony Prisk <linux@prisktech.co.nz>